Web.ArchLinux.API

Description

Arch Linux official repositories APIs and AUR APIs.

These two kinds of APIs are distinguished by APIClient. Functions over them return APIClient parametrize by corresponding APIType.

Overall, there are five APIs available, refer to https://wiki.archlinux.org/index.php/Official_repositories_web_interface and https://wiki.archlinux.org/index.php/Aurweb_RPC_interface.

API Client

newtype APIClient (k :: APIType) a

A wrapper of ClientM, with BaseUrl reflected to type level phantom APIType.

Constructors

APIClient

Fields

unWrapClientM :: ClientM a

data APIType

Two types of APIs.

Constructors

ArchLinux

Aur

class HasBaseUrl (k :: APIType) where

Class to reify BaseUrl from APIType.

Methods

getBaseUrl :: BaseUrl

runAPIClient :: forall s a. HasBaseUrl s => Manager -> APIClient s a -> IO (Either ClientError a)

Runs APIClient.

It calls getBaseUrl, then creates ClientEnv, finally calls runClientM.

runAPIClient' :: HasBaseUrl s => APIClient s a -> IO (Either ClientError a)

Like runAPIClient, but creates a Manager.

Arch Linux official

data SearchOptions

Options available in searching packages in Arch Linux official repositories.

See searchPackage.

Constructors

SearchOptions

Fields

_nameOrDescription :: Maybe Text

_exactName :: Maybe Text

_targetDescription :: Maybe Text

_targetRepositories :: [Repo]

_targetArchitectures :: [Arch]

_targetMaintianer :: Maybe Text

_targetPackager :: Maybe Text

_isFlagged :: Maybe Flagged

emptySearchOptions :: SearchOptions

An empty options value for convenient.

For example,

let options =
      emptySearchOptions
        & nameOrDescription ?~ "kea"
        & targetRepositories .~ [Community, CommunityTesting]
searchPackage options

searchs packages whose names or descriptions contain kea, from Community or Community-Testing.

getPackageDetails

Arguments

:: Repo
official repository

-> Arch
arch

-> Text
exact name

-> APIClient 'ArchLinux PackageInformation

Gets details of an exact package.

getPackageFiles

Arguments

:: Repo
official repository

-> Arch
arch

-> Text
exact name

-> APIClient 'ArchLinux PackageFiles

Gets files list of an exact package.

searchPackage :: SearchOptions -> APIClient 'ArchLinux (ArchLinuxResponse PackageInformation)

Searches packages.

See SearchOptions and emptySearchOptions.

AUR

data AurSearchType

Searches packages in AUR by what?

Constructors

ByName

ByNameOrDesc

ByMaintainer

ByDepends

ByMakedepends

ByOptdepends

ByCheckdepends

searchAur

Arguments

:: AurSearchType
search type

-> Text
search argument

-> APIClient 'Aur (AurResponse [AurSearch])

Searches packages in AUR.

getAurInfo

Arguments

:: [Text]
exact names

-> APIClient 'Aur (AurResponse [AurInfo])

Gets details of a set of packages in AUR.