Web.ArchLinux.Types

Description

This module defines types and their serializations used in API.

Fields' names are prefixed with _ for lenses generation, consider using Web.ArchLinux.Types.Lens to access data types smoothly.

Arch Linux official

data Repo

Official repositories.

Constructors

Core

Extra

Testing

Multilib

MultilibTesting

Community

CommunityTesting

KDEUnstable

data Arch

Official architectures.

Constructors

Any

I686

X86_64

data License

Liceses defined in https://archlinux.org/packages/core/any/licenses/.

Constructors

AGPL3

Apache

Artistic2_0

CDDL

CPL

EPL

FDL1_2

FDL1_3

GPL2

GPL3

LGPL2_1

LGPL3

LPPL

MPL

MPL2

PHP

PSF

PerlArtistic

RUBY

Unlicense

W3C

ZPL

Custom Text

data PackageInformation

Package details returned by getPackageDetails.

Constructors

PackageInformation

Fields

_pkgname :: Text

_pkgbase :: Text

_repo :: Repo

_arch :: Arch

_pkgver :: Text

_pkgrel :: Text

_epoch :: Int

_pkgdesc :: Text

_url :: Text

_filename :: Text

_compressedSize :: Int

_installedSize :: Int

_buildDate :: UTCTime

_lastUpdate :: UTCTime

_flageDate :: Maybe UTCTime

_maintainers :: [Text]

_packager :: Text

_groups :: [Text]

_licenses :: [License]

_conflicts :: [Text]

_provides :: [Text]

_replaces :: [Text]

_depends :: [Text]

_optdepends :: [Text]

_makedepends :: [Text]

_checkdepends :: [Text]

data PackageFiles

Package files list returned by getPackageFiles

Constructors

PackageFiles

Fields

_pkgname :: Text

_repo :: Repo

_arch :: Arch

_pkgLastUpdate :: UTCTime

_filesLastUpdate :: UTCTime

_filesCount :: Int

_dirCount :: Int

_files :: [FilePath]

data Flagged

Flagged package means out-of-date.

Constructors

Flagged

NotFlagged

data ArchLinuxResponse a

Response data type of searchPackage.

Constructors

ArchLinuxResponse

Fields

_version :: Int

_limit :: Int

_valid :: Bool

_results :: [a]

AUR

data AurSearch

Search results returned by searchAur.

Some of fields are renamed in this record type, for sharing overloaded lenses between data type returned by Arch Linux official API.

Constructors

AurSearch

Fields

_id :: Int

_name :: Text

_packageBaseID :: Int

_packageBase :: Text

_version :: Text

_description :: Maybe Text

_url :: Maybe Text

URL

_numVotes :: Int

_popularity :: Double

_outOfDate :: Maybe Int

UTCTime

_maintainer :: Maybe Text

_firstSubmitted :: Int

UTCTime

_lastModified :: Int

UTCTime

_urlPath :: Text

data AurInfo

Package details returned by getAurInfo.

This data type extends AurSearch informally, so it includes $sel:_search:AurInfo as a member.

Constructors

AurInfo

Fields

_search :: AurSearch

ID, Name, PackageBaseID, ...

_depends :: [Text]

Depends

_makedepends :: [Text]

MakeDepends

_optdepends :: [Text]

OptDepends

_checkdepends :: [Text]

CheckDepends

_conflicts :: [Text]

_provides :: [Text]

_replaces :: [Text]

_groups :: [Text]

_licenses :: [License]

_keywords :: [Text]

data AurResponseType

Return types of AUR API.

Constructors

Search

Multiinfo

Error

data AurResponse a

Response data type of AUR API.

Constructors

AurResponse

Fields

_version :: Int

_aurType :: AurResponseType

type

_resultCount :: Int

resultcount

_results :: a

_error :: Maybe Text

Available when aurType equals to Error.